[Refactor-792] Split Rootstore - not ready
