+++ src/App.tsx
@@ -10,6 +10,8 @@ import HomePageRedirect from './pages/HomePageRedirect';
 import { useTranslation } from 'react-i18next';
 import { useTheme } from '@material-ui/core/styles';
 import PopUpRedirect from './components/atoms/PopUpRedirect';
+import settingsStore from 'store/settings.store';
+
 // main components height - must add up to 100
 const headerHeight = '5vh';
 const pageContentHeight = '88vh';
